got rid of lots of warnings

+++ graph_query.cc
@@ -427,14 +427,14 @@ std::map<int, std::map<int, std::string>> does_trace_satisfy_conditions(
 ) {
     // isomap_index_to_node_index_to_span_id -> ii_to_ni_to_si
     std::vector<std::map<int, std::map<int, std::string>>> ii_to_ni_to_si_data_for_all_conditions;
-    for (int curr_cond_ind = 0; curr_cond_ind < conditions.size(); curr_cond_ind++) {
+    for (uint64_t curr_cond_ind = 0; curr_cond_ind < conditions.size(); curr_cond_ind++) {
         ii_to_ni_to_si_data_for_all_conditions.push_back(
             get_iso_maps_indices_for_which_trace_satifies_curr_condition(
                 trace_id, object_name, conditions, curr_cond_ind, evaluation_data, structural_results, ret));
     }
 
     std::map<int, std::map<int, std::string>> aggregate_result;
-    std::map<int, int> iso_map_to_satisfied_conditions_map;
+    std::map<int, uint64_t> iso_map_to_satisfied_conditions_map;
     for (auto vec_ele : ii_to_ni_to_si_data_for_all_conditions) {
         for (auto ele : vec_ele) {
             auto iso_map_index = ele.first;
